chore: trim test file and fix naming convention
Rename test_4106-fix-core-misc.py → test_4106_fix_core_misc.py (underscore
required by validate-test-names.py). Reduce 16 tests to 8: one focused
regression per bug, removing near-duplicate and internal-structure-probing
cases. Merge prettyprint width and splicing checks into one test.
